Add codesign wrapper script
The newly added `scripts/codesign` is a convenience wrapper around
codesign for easier and proper codesigning of Mach-O thin executables
like yabai. It expects two arguments: a path to the binary (e.g.,
`yabai`) and, if multiple codesigning identities are available, the
index of the listed codesigning identities.

Proper use of `codesign` is hard, and this script should alleviate some
of the trouble users have been having, especially with regards to the
changes in recent versions of macOS.