Fix window resizing not always updating layout
Fixes #1199: resizing a window didn't always update
the layout split. The underlying reason is that the
mouse down coordinates for window resizing are not
always within the `CGRect` of the window's frame;
often they are just outside the window's frame when
the resize cursor appears.

I changed the event tap location to `kCGAnnotatedSessionEventTap`,
which includes the field `kCGMouseEventWindowUnderMousePointerThatCanHandleThisEvent`
in the event metadata that always accurately identifies the window
that the click affects. This **does** alter the point in the input
stack at which Yabai observes events, but examining `EVENT_TAP_CALLBACK(mouse_handler)`
and user-testing the mouse features seems to show that changing
the observation point does not cause any regression.