Reorder resize vs move based on size delta
When setting the size of a window via `window_manager_set_window_frame`,
either resize before moving, if the window is getting smaller, or
move before resizing, if the window is getting bigger. This avoids
window size conflicts with the system window manager boundaries, like
display edge. If a too-big window is moved to the edge of the screen,
the system display manager will force it to be resized to fit within
the screen dimensions, and the `AXUIElementSetAttributeValue` API call
issued by yabai to resize it gets ignored.