Minimized windows should not be managed
Add a check to exclude minimized windows in
`window_manager_should_manage_window()`, because minimized windows
should not be added to the view. The unminimized event handler will add
them to the view when they become visible.

This also removes a check from one call site of
`window_manager_should_manage_window()` that also checked whether the
window is minimized.