keep layer unchanged during untiling
When untiling a window, its layer is set to normal which has the effect
of it also becoming a "topmost" window, since the default seems to be
"below" for all other windows. Resulting in issues like #1929

My suggestion is to not alter the layer during untiling, requiring to
set it explicitely.

(fwiw: my initial attempt at changing the default layer to `normal`
 didn't work as I expected, but maybe that'd be a solution)